Mijozlar, Qarzdorlar, Savdolar bo'limida qarzdor clientni summasi aniq ko'rinadi, vozvrat qilingandan keyin ham. Hisobotlar bo'limida qarzga olingan tovarlar ham ko'rinadi

+++ MarketSystem.Client/lib/features/sales/data/datasources/sale_remote_data_source.dart
@@ -35,7 +35,7 @@ class SaleRemoteDataSource {
   Future<dynamic> addSaleItem({
     required String saleId,
     required String productId,
-    required int quantity,
+    required double quantity,  // ✅ DECIMAL
     required double salePrice,
     required double minSalePrice,
     String? comment,
